Fixed exception printing
